05/10 커밋/////

+++ teamproject1/src/main/java/com/example/domain/BoardVO.java
@@ -13,6 +13,8 @@ public class BoardVO {
 	private String writer;
 	private Date regdate;
 	private Date updatedate;
-	private int hit;
+	private Long hit;
 	private String nickname;
+	
+
 }
